tasks/computer-vision/image-generation/art_generation/requirements.txt
tensorflow>=1.14.0
mlflow>=1.2.0
scipy==1.1.0
matplotlib>=3.1.1
pillow>=6.1.0